Refactor ecocredits data fetch to custom hook
